Edge/uploadFileToBlob/requirements.txt
azure-storage-blob==12.14.1
python-dotenv==0.20.0